requirements.txt
python-telegram-bot
httpx
Pillow>=10.0
